+++ apps/ranking-service/internal/domain/ranking/entity.go
@@ -2,6 +2,6 @@ package ranking
 
 type RankEntry struct {
 	UserID string
-	Score  float64
+	Score  uint64
 	Rank   int64 // 1-based
 }
