+++ apps/score-service/internal/application/score/port.go
@@ -3,5 +3,5 @@ package score
 import "context"
 
 type UseCase interface {
-	SubmitScore(ctx context.Context, userID string, score float64) error
+	SubmitScore(ctx context.Context, userID string, score uint64) error
 }
